Move im2col.cc and im2col.cu to header files
