+++ src/operator/convolution.cu
@@ -10,7 +10,7 @@
 #if MXNET_USE_CUDNN == 1
 #include "./cudnn_convolution-inl.h"
 #endif  // MXNET_USE_CUDNN
-#include "./nn/im2col.h"
+#include "./nn/im2col.cuh"
 
 namespace mxnet {
 namespace op {
