--- src/operator/nn/im2col.cc
@@ -1,320 +1,0 @@
-/*!
- ******************* BEGIN Caffe Copyright Notice and Disclaimer ****************
- *
- * COPYRIGHT
- * 
- * All contributions by the University of California:
- * Copyright (c) 2014-2017 The Regents of the University of California (Regents)
- * All rights reserved.
- * 
- * All other contributions:
- * Copyright (c) 2014-2017, the respective contributors
- * All rights reserved.
- * 
- * Caffe uses a shared copyright model: each contributor holds copyright over
- * their contributions to Caffe. The project versioning records all such
- * contribution and copyright details. If a contributor wants to further mark
- * their specific copyright on a particular contribution, they should indicate
- * their copyright solely in the commit message of the change when it is
- * committed.
- * 
- * LICENSE
- * 
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met: 
- * 
- * 1. Redistributions of source code must retain the above copyright notice, this
- * list of conditions and the following disclaimer. 
- * 2. Redistributions in binary form must reproduce the above copyright notice,
- * this list of conditions and the following disclaimer in the documentation
- * and/or other materials provided with the distribution. 
- * 
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
- * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
- * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
- * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
- * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
- * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
- * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
- * ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
- * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- * 
- * CONTRIBUTION AGREEMENT
- * 
- * By contributing to the BVLC/caffe repository through pull-request, comment,
- * or otherwise, the contributor releases their content to the
- * license and copyright terms herein.
- *
- ***************** END Caffe Copyright Notice and Disclaimer ********************
- *
- * Copyright (c) 2017 by Contributors
- * \file im2col.cc
- * \brief Function definitions of converting an image to
- * column matrix based on kernel, padding, and dilation.
- * These functions are mainly used in convolution operators.
- * The implementation of the im2col and col2im algorithms
- * are copied from Caffe with minor interface modifications
- * adapting to MXNet data structures.
- */
-
-#include "./im2col.h"
-#include <vector>
-
-using std::vector;
-
-namespace mxnet {
-namespace op {
-
-// Function uses casting from int to unsigned to compare if value of
-// parameter a is greater or equal to zero and lower than value of
-// parameter b. The b parameter is of type signed and is always positive,
-// therefore its value is always lower than 0x800... where casting
-// negative value of a parameter converts it to value higher than 0x800...
-// The casting allows to use one condition instead of two.
-inline bool is_a_ge_zero_and_a_lt_b(int a, int b) {
-  return static_cast<unsigned>(a) < static_cast<unsigned>(b);
-}
-
-template <typename DType>
-void im2col_cpu(const DType* data_im, const int channels,
-    const int height, const int width, const int kernel_h, const int kernel_w,
-    const int pad_h, const int pad_w,
-    const int stride_h, const int stride_w,
-    const int dilation_h, const int dilation_w,
-    DType* data_col) {
-  const int output_h = (height + 2 * pad_h -
-    (dilation_h * (kernel_h - 1) + 1)) / stride_h + 1;
-  const int output_w = (width + 2 * pad_w -
-    (dilation_w * (kernel_w - 1) + 1)) / stride_w + 1;
-  const int channel_size = height * width;
-  for (int channel = channels; channel--; data_im += channel_size) {
-    for (int kernel_row = 0; kernel_row < kernel_h; kernel_row++) {
-      for (int kernel_col = 0; kernel_col < kernel_w; kernel_col++) {
-        int input_row = -pad_h + kernel_row * dilation_h;
-        for (int output_rows = output_h; output_rows; output_rows--) {
-          if (!is_a_ge_zero_and_a_lt_b(input_row, height)) {
-            for (int output_cols = output_w; output_cols; output_cols--) {
-              *(data_col++) = 0;
-            }
-          } else {
-            int input_col = -pad_w + kernel_col * dilation_w;
-            for (int output_col = output_w; output_col; output_col--) {
-              if (is_a_ge_zero_and_a_lt_b(input_col, width)) {
-                *(data_col++) = data_im[input_row * width + input_col];
-              } else {
-                *(data_col++) = 0;
-              }
-              input_col += stride_w;
-            }
-          }
-          input_row += stride_h;
-        }
-      }
-    }
-  }
-}
-
-// Explicit instantiation
-template void im2col_cpu<float>(const float* data_im, const int channels,
-    const int height, const int width, const int kernel_h, const int kernel_w,
-    const int pad_h, const int pad_w, const int stride_h,
-    const int stride_w, const int dilation_h, const int dilation_w,
-    float* data_col);
-template void im2col_cpu<double>(const double* data_im, const int channels,
-    const int height, const int width, const int kernel_h, const int kernel_w,
-    const int pad_h, const int pad_w, const int stride_h,
-    const int stride_w, const int dilation_h, const int dilation_w,
-    double* data_col);
-template void im2col_cpu<mshadow::half::half_t>(const mshadow::half::half_t* data_im,
-    const int channels, const int height, const int width, const int kernel_h,
-    const int kernel_w, const int pad_h, const int pad_w, const int stride_h,
-    const int stride_w, const int dilation_h, const int dilation_w,
-    mshadow::half::half_t* data_col);
-
-template <typename DType>
-inline void im2col_nd_core_cpu(const DType* data_input, const bool im2col,
-    const int num_spatial_axes, const int* im_shape, const int* col_shape,
-    const int* kernel_shape, const int* pad, const int* stride,
-    const int* dilation, DType* data_output, OpReqType req = mxnet::kWriteTo) {
-  if (mxnet::kNullOp == req) return;
-  if (!im2col) {
-    int im_size = im_shape[0];
-    for (int i = 0; i < num_spatial_axes; ++i) {
-      im_size *= im_shape[1 + i];
-    }
-    if (mxnet::kAddTo != req) {
-      fill_array(im_size, DType(0), data_output);
-    }
-  }
-  int kernel_size = 1;
-  for (int i = 0; i < num_spatial_axes; ++i) {
-    kernel_size *= kernel_shape[i];
-  }
-  const int channels_col = col_shape[0];
-  vector<int> d_offset(num_spatial_axes, 0);
-  vector<int> d_iter(num_spatial_axes, 0);
-  for (int c_col = 0; c_col < channels_col; ++c_col) {
-    // Loop over spatial axes in reverse order to compute a per-axis offset.
-    int offset = c_col;
-    for (int d_i = num_spatial_axes - 1; d_i >= 0; --d_i) {
-      if (d_i < num_spatial_axes - 1) {
-        offset /= kernel_shape[d_i + 1];
-      }
-      d_offset[d_i] = offset % kernel_shape[d_i];
-    }
-    for (bool incremented = true; incremented; ) {
-      // Loop over spatial axes in forward order to compute the indices in the
-      // image and column, and whether the index lies in the padding.
-      int index_col = c_col;
-      int index_im = c_col / kernel_size;
-      bool is_padding = false;
-      for (int d_i = 0; d_i < num_spatial_axes; ++d_i) {
-        const int d = d_iter[d_i];
-        const int d_im = d * stride[d_i] - pad[d_i] +
-            d_offset[d_i] * dilation[d_i];
-        is_padding |= d_im < 0 || d_im >= im_shape[d_i + 1];
-        index_col *= col_shape[d_i + 1];
-        index_col += d;
-        index_im *= im_shape[d_i + 1];
-        index_im += d_im;
-      }
-      if (im2col) {
-        if (is_padding) {
-          data_output[index_col] = 0;
-        } else {
-          data_output[index_col] = data_input[index_im];
-        }
-      } else if (!is_padding) {  // col2im
-        data_output[index_im] += data_input[index_col];
-      }
-      // Loop over spatial axes in reverse order to choose an index,
-      // like counting.
-      incremented = false;
-      for (int d_i = num_spatial_axes - 1; d_i >= 0; --d_i) {
-        const int d_max = col_shape[d_i + 1];
-        CHECK_LT(d_iter[d_i], d_max);
-        if (d_iter[d_i] == d_max - 1) {
-          d_iter[d_i] = 0;
-        } else {  // d_iter[d_i] < d_max - 1
-          ++d_iter[d_i];
-          incremented = true;
-          break;
-        }
-      }
-    }  // while(incremented) {
-  }  // for (int c = 0; c < channels_col; ++c) {
-}
-
-template <typename DType>
-void im2col_nd_cpu(const DType* data_im, const int num_spatial_axes,
-    const int* im_shape, const int* col_shape,
-    const int* kernel_shape, const int* pad, const int* stride,
-    const int* dilation, DType* data_col) {
-  const bool kIm2Col = true;
-  im2col_nd_core_cpu(data_im, kIm2Col, num_spatial_axes, im_shape, col_shape,
-                  kernel_shape, pad, stride, dilation, data_col);
-}
-
-// Explicit instantiation
-template void im2col_nd_cpu<float>(const float* data_im,
-    const int num_spatial_axes,
-    const int* im_shape, const int* col_shape,
-    const int* kernel_shape, const int* pad, const int* stride,
-    const int* dilation, float* data_col);
-template void im2col_nd_cpu<double>(const double* data_im,
-    const int num_spatial_axes,
-    const int* im_shape, const int* col_shape,
-    const int* kernel_shape, const int* pad, const int* stride,
-    const int* dilation, double* data_col);
-template void im2col_nd_cpu<mshadow::half::half_t>(const mshadow::half::half_t* data_im,
-    const int num_spatial_axes,
-    const int* im_shape, const int* col_shape,
-    const int* kernel_shape, const int* pad, const int* stride,
-    const int* dilation, mshadow::half::half_t* data_col);
-
-template <typename DType>
-void col2im_cpu(const DType* data_col, const int channels,
-    const int height, const int width, const int kernel_h, const int kernel_w,
-    const int pad_h, const int pad_w,
-    const int stride_h, const int stride_w,
-    const int dilation_h, const int dilation_w,
-    DType* data_im, OpReqType req) {
-  if (mxnet::kNullOp == req) return;
-  if (mxnet::kAddTo != req) {
-    fill_array(height * width * channels, static_cast<DType>(0), data_im);
-  }
-  const int output_h = (height + 2 * pad_h -
-    (dilation_h * (kernel_h - 1) + 1)) / stride_h + 1;
-  const int output_w = (width + 2 * pad_w -
-    (dilation_w * (kernel_w - 1) + 1)) / stride_w + 1;
-  const int channel_size = height * width;
-  for (int channel = channels; channel--; data_im += channel_size) {
-    for (int kernel_row = 0; kernel_row < kernel_h; kernel_row++) {
-      for (int kernel_col = 0; kernel_col < kernel_w; kernel_col++) {
-        int input_row = -pad_h + kernel_row * dilation_h;
-        for (int output_rows = output_h; output_rows; output_rows--) {
-          if (!is_a_ge_zero_and_a_lt_b(input_row, height)) {
-            data_col += output_w;
-          } else {
-            int input_col = -pad_w + kernel_col * dilation_w;
-            for (int output_col = output_w; output_col; output_col--) {
-              if (is_a_ge_zero_and_a_lt_b(input_col, width)) {
-                data_im[input_row * width + input_col] += *data_col;
-              }
-              data_col++;
-              input_col += stride_w;
-            }
-          }
-          input_row += stride_h;
-        }
-      }
-    }
-  }
-}
-
-// Explicit instantiation
-template void col2im_cpu<float>(const float* data_col, const int channels,
-    const int height, const int width, const int kernel_h, const int kernel_w,
-    const int pad_h, const int pad_w, const int stride_h,
-    const int stride_w, const int dilation_h, const int dilation_w,
-    float* data_im, OpReqType req);
-template void col2im_cpu<double>(const double* data_col, const int channels,
-    const int height, const int width, const int kernel_h, const int kernel_w,
-    const int pad_h, const int pad_w, const int stride_h,
-    const int stride_w, const int dilation_h, const int dilation_w,
-    double* data_im, OpReqType req);
-template void col2im_cpu<mshadow::half::half_t>(const mshadow::half::half_t* data_col,
-    const int channels, const int height, const int width, const int kernel_h, const int kernel_w,
-    const int pad_h, const int pad_w, const int stride_h,
-    const int stride_w, const int dilation_h, const int dilation_w,
-    mshadow::half::half_t* data_im, OpReqType req);
-
-template <typename DType>
-void col2im_nd_cpu(const DType* data_col, const int num_spatial_axes,
-    const int* im_shape, const int* col_shape,
-    const int* kernel_shape, const int* pad, const int* stride,
-    const int* dilation, DType* data_im, OpReqType req) {
-  const bool kIm2Col = false;
-  im2col_nd_core_cpu(data_col, kIm2Col, num_spatial_axes, im_shape, col_shape,
-                     kernel_shape, pad, stride, dilation, data_im, req);
-}
-
-// Explicit instantiation
-template void col2im_nd_cpu<float>(const float* data_col,
-    const int num_spatial_axes,
-    const int* im_shape, const int* col_shape,
-    const int* kernel_shape, const int* pad, const int* stride,
-    const int* dilation, float* data_im, OpReqType req);
-template void col2im_nd_cpu<double>(const double* data_col,
-    const int num_spatial_axes,
-    const int* im_shape, const int* col_shape,
-    const int* kernel_shape, const int* pad, const int* stride,
-    const int* dilation, double* data_im, OpReqType req);
-template void col2im_nd_cpu<mshadow::half::half_t>(const mshadow::half::half_t* data_col,
-    const int num_spatial_axes,
-    const int* im_shape, const int* col_shape,
-    const int* kernel_shape, const int* pad, const int* stride,
-    const int* dilation, mshadow::half::half_t* data_im, OpReqType req);
-}  // namespace op
-}  // namespace mxnet
